reorganize python-lib folders
