+++ python-lib/nlp/ontology_tagging/formatting/formatter_by_document.py
@@ -9,7 +9,7 @@ from collections import defaultdict
 
 from spacy.tokens import Span, Doc
 
-from utils.plugin_io_utils import unique_list, generate_unique_columns
+from utils.cleaning_utils import unique_list, generate_unique_columns
 from nlp.utils import get_span_text, unicode_normalize_text
 
 from .base import FormatterBase
